backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.1.5
gunicorn==20.1.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
Pillow==10.0.0